requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
requests==2.31.0
gunicorn==21.2.0